Extract HiveAuthenticationConfig
